Inline useUrlConfig into useSearch
